Move validation check inside CheckoutProcessor's processCheckout
