+++ assets/js/base/context/cart-checkout/checkout/processor/index.js
@@ -8,23 +8,35 @@ import {
 	useShippingDataContext,
 	useBillingDataContext,
 	usePaymentMethodDataContext,
+	useValidationContext,
 } from '@woocommerce/base-context';
 import { useEffect, useRef, useCallback } from '@wordpress/element';
 import { useStoreNotices } from '@woocommerce/base-hooks';
+import withScrollToTop from '@woocommerce/base-hocs/with-scroll-to-top';
 
 /**
  * CheckoutProcessor component. @todo Needs to consume all contexts.
  *
  * Subscribes to checkout context and triggers processing via the API.
  */
-const CheckoutProcessor = () => {
-	const { onCheckoutProcessing, dispatchActions } = useCheckoutContext();
+const CheckoutProcessor = ( { scrollToTop } ) => {
+	const {
+		onCheckoutProcessing,
+		onCheckoutCompleteError,
+		dispatchActions,
+	} = useCheckoutContext();
+	const {
+		hasValidationErrors,
+		showAllValidationErrors,
+	} = useValidationContext();
 	const { shippingAddress } = useShippingDataContext();
 	const { billingData } = useBillingDataContext();
 	const { activePaymentMethod } = usePaymentMethodDataContext();
 	const { addErrorNotice } = useStoreNotices();
 	const currentBillingData = useRef( billingData );
 	const currentShippingAddress = useRef( shippingAddress );
+
+	const withErrors = hasValidationErrors();
 
 	useEffect( () => {
 		currentBillingData.current = billingData;
@@ -37,6 +49,10 @@ const CheckoutProcessor = () => {
 	 * @return {boolean} True if everything was successful.
 	 */
 	const processCheckout = useCallback( async () => {
+		if ( withErrors ) {
+			return false;
+		}
+
 		await triggerFetch( {
 			path: '/wc/store/checkout',
 			method: 'POST',
@@ -95,6 +111,7 @@ const CheckoutProcessor = () => {
 		activePaymentMethod,
 		currentBillingData,
 		currentShippingAddress,
+		withErrors,
 	] );
 
 	/**
@@ -107,7 +124,17 @@ const CheckoutProcessor = () => {
 		};
 	}, [ onCheckoutProcessing, processCheckout ] );
 
+	useEffect( () => {
+		const unsubscribeCompleteError = onCheckoutCompleteError( () => {
+			showAllValidationErrors();
+			scrollToTop( { focusableSelector: 'input:invalid' } );
+		} );
+		return () => {
+			unsubscribeCompleteError();
+		};
+	}, [ onCheckoutCompleteError ] );
+
 	return null;
 };
 
-export default CheckoutProcessor;
+export default withScrollToTop( CheckoutProcessor );
